Separate from .model classes unchanged in 3.0 vs. 2.1
- These are listed in the "Summary of Major Changes and New
  Functionality", §3.3 Areas Unchanged from Version 2.1.
- Place these in new .model.common, .model.internationalstring.
- Allow them to be imported from sdmx.model without warning.
